% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.R
\name{s2_states}
\alias{s2_states}
\title{Get states geographic boundaries}
\usage{
s2_states(year = as.character(2010:2023))
}
\arguments{
\item{year}{character data year passed to tigris to get tract boundaries}
}
\description{
Get states geographic boundaries
}
\examples{
s2_states(year = "2020")
}
