% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.R
\name{s2_tracts}
\alias{s2_tracts}
\title{Get census tract geographic boundaries}
\usage{
s2_tracts(state, year = as.character(2010:2023))
}
\arguments{
\item{state}{character string of state number or abbreviation}

\item{year}{character data year passed to tigris to get tract boundaries}
}
\value{
a tibble of tracts with a s2_geography column
}
\description{
Get census tract geographic boundaries
}
\examples{
s2_tracts("OH", "2023")
}
